REFACTOR: Remove Tilt/ES6ModuleTranspiler
