--- lib/es6_module_transpiler/rails/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module ES6ModuleTranspiler
-  module Rails
-    VERSION = '0.4.0'
-  end
-end
