--- lib/es6_module_transpiler/tilt.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require 'tilt'
-require 'es6_module_transpiler/tilt/es6_module_transpiler_template'
-
-# Tilt.register Tilt::ES6ModuleTranspilerTemplate, 'es6'
